+++ opentelemetry-otlp/src/exporter/http/mod.rs
@@ -105,7 +105,7 @@ use opentelemetry_http::hyper::HyperClient;
 
 /// Configuration of the http transport
 #[derive(Debug, Default)]
-pub struct HttpConfig {
+pub(crate) struct HttpConfig {
     /// Select the HTTP client
     client: Option<Arc<dyn HttpClient>>,
 
@@ -729,7 +729,7 @@ fn add_header_from_string(input: &str, headers: &mut HashMap<HeaderName, HeaderV
 }
 
 /// Expose interface for modifying builder config.
-pub trait HasHttpConfig {
+pub(crate) trait HasHttpConfig {
     /// Return a mutable reference to the config within the exporter builders.
     fn http_client_config(&mut self) -> &mut HttpConfig;
 }
@@ -741,7 +741,7 @@ impl HasHttpConfig for HttpExporterBuilder {
     }
 }
 
-/// This trait will be implemented for every struct that implemented [`HasHttpConfig`] trait.
+/// Expose methods to override HTTP-specific configuration.
 ///
 /// ## Examples
 /// ```