+++ opentelemetry-otlp/src/logs.rs
@@ -11,10 +11,10 @@ use std::time;
 use crate::{exporter::HasExportConfig, ExporterBuildError, NoExporterBuilderSet};
 
 #[cfg(feature = "grpc-tonic")]
-use crate::{HasTonicConfig, TonicExporterBuilder, TonicExporterBuilderSet};
+use crate::{exporter::tonic::HasTonicConfig, TonicExporterBuilder, TonicExporterBuilderSet};
 
 #[cfg(any(feature = "http-proto", feature = "http-json"))]
-use crate::{HasHttpConfig, HttpExporterBuilder, HttpExporterBuilderSet};
+use crate::{exporter::http::HasHttpConfig, HttpExporterBuilder, HttpExporterBuilderSet};
 
 /// Compression algorithm to use, defaults to none.
 pub const OTEL_EXPORTER_OTLP_LOGS_COMPRESSION: &str = "OTEL_EXPORTER_OTLP_LOGS_COMPRESSION";
@@ -97,7 +97,7 @@ impl HasExportConfig for LogExporterBuilder<HttpExporterBuilderSet> {
 
 #[cfg(feature = "grpc-tonic")]
 impl HasTonicConfig for LogExporterBuilder<TonicExporterBuilderSet> {
-    fn tonic_config(&mut self) -> &mut crate::TonicConfig {
+    fn tonic_config(&mut self) -> &mut crate::exporter::tonic::TonicConfig {
         &mut self.client.0.tonic_config
     }
 }